requirements.txt
Flask==1.1.2
google-cloud-datastore==1.12.0
flask-bootstrap
google-cloud-bigquery==1.24.0